ZEN: logo: Remove random logos, reduce diff with upstream
To reduce maintenance cost of carrying extra logos with Zen, lets remove
unneccessary code changes whenever possible.  This includes refactoring
the addition of logos so the diff from upstream is as straight-forward
and unlikely to result in a merge conflict.

This also means we've dropped random logos, but I really don't know
anyone who uses this.  More of a gimmick rather than a feature.  Will
never get upstreamed.